requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
supabase==1.0.3
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0